Поправил модуль отзывов.

+++ app/Modules/Review/Actions/Admin/ReviewCreateAction.php
@@ -58,9 +58,9 @@ class ReviewCreateAction extends Action
     /**
      * Рейтинг.
      *
-     * @var float|null
+     * @var int|null
      */
-    public ?float $rating = null;
+    public ?int $rating = null;
 
     /**
      * Статус.
