+++ app/Modules/Review/Entities/Review.php
@@ -56,9 +56,9 @@ class Review extends Entity
     /**
      * Рейтинг.
      *
-     * @var float|null
+     * @var int|null
      */
-    public ?float $rating = null;
+    public ?int $rating = null;
 
     /**
      * Статус.
